Suggestion for Section 1 - Challenge Question 2
I added a solution that uses more string manipulation. Your solution isnt wrong but using string methods should run faster. For example, lets say you have a very long string like:
"thisisasuperlongstringwithoutspaces"

The request is to find and remove the first occurrence of 'c'. You would need to loop through almost the entire string till you find the character to remove. This is referred to O(n). The Big O notation is a formulaic expression to describe the complexity of an algorithm. The reason your solution is O(n) is that if the input was n-length long, the algorithm could run up to n iterations.